+++ ast_language/__init__.py
@@ -1,4 +1,5 @@
 from .ast_util import *
+from .columns_util import *
 from .linq_util import *
 from .parse import *
 from .transform import *
