+++ ast_language/linq_util.py
@@ -1,4 +1,4 @@
-from . import ast_util
+from .ast_util import unwrap_ast
 
 import ast
 
@@ -10,13 +10,13 @@ class Select(ast.AST):
         self.selector = selector
 
 
-class SelectTransformer(ast.NodeTransformer):
+class InsertLINQNodesTransformer(ast.NodeTransformer):
     def visit_Call(self, node):
         if isinstance(node.func, ast.Attribute) and node.func.attr == 'Select':
             if len(node.args) != 1:
                 raise SyntaxError('Select() call must have exactly one argument')
             if isinstance(node.args[0], ast.Str):
-                node.args[0] = ast_util.unwrap_ast(ast.parse(node.args[0].s))
+                node.args[0] = unwrap_ast(ast.parse(node.args[0].s))
             if not isinstance(node.args[0], ast.Lambda):
                 raise SyntaxError('Select() call argument must be a lambda')
             return Select(source=node.func.value, selector=node.args[0])
@@ -24,5 +24,13 @@ class SelectTransformer(ast.NodeTransformer):
             return self.generic_visit(node)
 
 
-def transform_selects(node):
-    return SelectTransformer().visit(node)
+def insert_linq_nodes(python_ast):
+    return InsertLINQNodesTransformer().visit(python_ast)
+
+
+class RemoveLINQNodesTransformer(ast.NodeTransformer):
+    pass
+
+
+def remove_linq_nodes(python_ast):
+    return RemoveLINQNodesTransformer().visit(python_ast)
